♻️ [Refactor]: study room component 이동
