+++ src/entities/study/curriculum/ui/Curriculum.tsx
