+++ src/entities/study/curriculum/ui/CurriculumCard.tsx
