+++ src/entities/study/join/ui/InfoModal.tsx
