+++ src/entities/study/join/ui/ManageCard.tsx
