+++ src/entities/study/overview/api/rules.api.types.ts
