+++ src/entities/study/overview/model/rules.queries.ts
@@ -1,10 +1,10 @@
-import { useQuery, useMutation, useQueryClient } from "@tanstack/react-query";
+import { useMutation, useQuery, useQueryClient } from "@tanstack/react-query";
 
 import {
   EditRuleRequest,
   editRules,
   getRules,
-} from "@/entities/study/rule/api/rules.api";
+} from "@/entities/study/overview/api/rules.api";
 
 export const RulesQueryKey = (studyId: number) => ["study", "rules", studyId];
