+++ src/entities/study/overview/ui/ContentInputBox.tsx
