+++ src/entities/study/overview/ui/ManageContentCard.tsx
