+++ src/entities/study/overview/ui/ManageOverviewCard.tsx
