+++ src/entities/study/overview/ui/PreviewBox.tsx
