+++ src/entities/study/schedule/api/schedule.api.ts
@@ -1,19 +1,19 @@
 export type {
-  EditRuleResponse,
   EditRuleRequest,
+  EditRuleResponse,
   EditStudyRule,
-  StudyRule,
   GetStudyRulesResponse,
-} from "@/entities/study/rule/api/rules.api.types";
+  StudyRule,
+} from "@/entities/study/overview/api/rules.api.types";
 
 import { createJsonRequestInit, fetchAPI } from "@/shared/api/util/fetcher";
 
-import { API_ENDPOINT } from "@/shared/api/lib";
 import {
   GetSchedulesResponse,
   PostStudyScheduleResponse,
 } from "@/entities/study/schedule/api/schedule.api.types";
 import { GetStudySchedule } from "@/features/study-room/types/study-room.api";
+import { API_ENDPOINT } from "@/shared/api/lib";
 
 // 스터디 전체 일정 조회
 export async function getSchedules(): Promise<GetSchedulesResponse[]> {
