+++ src/entities/study/schedule/ui/ScheduleModal.tsx
