+++ src/entities/study/setting/ui/SettingCard.tsx
